Replace Debug{Before/After}Transition with unified DebugTransition
