+++ kore/kore.cabal
@@ -328,14 +328,13 @@ library
     Kore.Internal.TermLike.TermLike
     Kore.Internal.Variable
     Kore.Log
-    Kore.Log.DebugAfterTransition
     Kore.Log.DebugAppliedRewriteRules
-    Kore.Log.DebugBeforeTransition
     Kore.Log.DebugBeginClaim
     Kore.Log.DebugEvaluateCondition
     Kore.Log.DebugProven
     Kore.Log.DebugSolver
     Kore.Log.DebugSubstitutionSimplifier
+    Kore.Log.DebugTransition
     Kore.Log.DebugUnification
     Kore.Log.ErrorBottomTotalFunction
     Kore.Log.ErrorDecidePredicateUnknown
