+++ kore/src/Kore/Reachability/Prove.hs
@@ -76,11 +76,11 @@ import Kore.Internal.Predicate (
     pattern PredicateCeil,
     pattern PredicateNot,
  )
-import Kore.Log.DebugAfterTransition (
+import Kore.Log.DebugTransition (
+    debugBeforeTransition,
     debugAfterTransition,
     debugFinalTransition,
  )
-import Kore.Log.DebugBeforeTransition
 import Kore.Log.DebugBeginClaim
 import Kore.Log.DebugProven
 import Kore.Log.InfoExecBreadth
